fix(web): address follow-up message comment review
